// webapp/src/app/core/api/v1/model/applicationElementResponse.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApplicationElementResponse { 
    applicationId?: number;
    elementId?: number;
    rowId?: number;
    currencyResponse?: number | null;
    textResponse?: string | null;
    longTextResponse?: string | null;
    intResponse?: number | null;
    bitResponse?: boolean | null;
    dateResponse?: string | null;
}